libchiaki.pxd:  Restructure in preparation for using chiaki4deck instead of chiaki
Despite the name, chiaki4deck maintains support for other platforms, and unlike chiaki itself, is actually maintained.

Unfortunately, the library part of both projects does not have a stable API, so tracking it is going to be...  fun...

After learning more about Cython, our "custom" header file wasn't necessary - it was more of a hack for autopxd.

Now that we likely won't be using autopxd again for this project, we don't need that workaround.

Next phase:  Update all of the structs and function prototypes that changed...  :(